requirements.txt
dj-database-url==1.2.0
Django==4.1.3
psycopg2==2.9.5
python-decouple==3.8
virtualenv==20.21.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
django-filter==22.1
drf-yasg==1.21.5